requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.6
oauthlib==3.1.0
pylint==2.6.0
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
tweepy==3.9.0
urllib3==1.25.10
webdriver-manager==3.2.2
wrapt==1.12.1
